extract dataclasses for lulucf2030
