+++ generatorcore/lulucf2030/__init__.py
@@ -1,90 +1,7 @@
-from dataclasses import dataclass, field
 from ..inputs import Inputs
 from ..utils import div
 from .. import lulucf2018
-
-
-@dataclass
-class LColVars2030:
-    area_ha: float = None  # type: ignore
-    CO2e_production_based_per_t: float = None  # type: ignore
-    pct_x: float = None  # type: ignore
-    CO2e_production_based_per_MWh: float = None  # type: ignore
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    demand_change: float = None  # type: ignore
-    area_ha_change: float = None  # type: ignore
-    CO2e_combustion_based_per_t: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-    emplo_existing: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    area_ha_available_pct_of_action: float = None  # type: ignore
-    area_ha_available: float = None  # type: ignore
-    change_within_category: float = None  # type: ignore
-    change_wet_org_low: float = None  # type: ignore
-    change_wet_org_high: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-
-
-@dataclass
-class L30:
-    l: LColVars2030 = field(default_factory=LColVars2030)
-    g: LColVars2030 = field(default_factory=LColVars2030)
-    g_forest: LColVars2030 = field(default_factory=LColVars2030)
-    g_forest_managed: LColVars2030 = field(default_factory=LColVars2030)
-    g_forest_natural: LColVars2030 = field(default_factory=LColVars2030)
-    g_crop: LColVars2030 = field(default_factory=LColVars2030)
-    g_crop_org: LColVars2030 = field(default_factory=LColVars2030)
-    g_crop_min_conv: LColVars2030 = field(default_factory=LColVars2030)
-    g_crop_min_hum: LColVars2030 = field(default_factory=LColVars2030)
-    g_crop_org_low: LColVars2030 = field(default_factory=LColVars2030)
-    g_crop_org_high: LColVars2030 = field(default_factory=LColVars2030)
-    g_grass: LColVars2030 = field(default_factory=LColVars2030)
-    g_grass_min_conv: LColVars2030 = field(default_factory=LColVars2030)
-    g_grass_org_low: LColVars2030 = field(default_factory=LColVars2030)
-    g_grass_org_high: LColVars2030 = field(default_factory=LColVars2030)
-    g_grove: LColVars2030 = field(default_factory=LColVars2030)
-    g_grove_min: LColVars2030 = field(default_factory=LColVars2030)
-    g_grove_org: LColVars2030 = field(default_factory=LColVars2030)
-    g_grove_org_low: LColVars2030 = field(default_factory=LColVars2030)
-    g_grove_org_high: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_min: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_org_low: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_org_high: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_org_low_r: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_org_low_rp: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_org_high_r: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_org_high_rp: LColVars2030 = field(default_factory=LColVars2030)
-    g_water: LColVars2030 = field(default_factory=LColVars2030)
-    g_water_min: LColVars2030 = field(default_factory=LColVars2030)
-    g_water_org: LColVars2030 = field(default_factory=LColVars2030)
-    g_settlement: LColVars2030 = field(default_factory=LColVars2030)
-    g_settlement_org: LColVars2030 = field(default_factory=LColVars2030)
-    g_settlement_min: LColVars2030 = field(default_factory=LColVars2030)
-    g_settlement_org_low: LColVars2030 = field(default_factory=LColVars2030)
-    g_settlement_org_high: LColVars2030 = field(default_factory=LColVars2030)
-    g_other: LColVars2030 = field(default_factory=LColVars2030)
-    g_wood: LColVars2030 = field(default_factory=LColVars2030)
-    pyr: LColVars2030 = field(default_factory=LColVars2030)
-    g_planning: LColVars2030 = field(default_factory=LColVars2030)
-    g_grass_org: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_org: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_org_r: LColVars2030 = field(default_factory=LColVars2030)
-    g_wet_org_rp: LColVars2030 = field(default_factory=LColVars2030)
-    g_water_org_low: LColVars2030 = field(default_factory=LColVars2030)
-    g_water_org_high: LColVars2030 = field(default_factory=LColVars2030)
+from .l30 import L30
 
 
 def calc(inputs: Inputs, *, l18: lulucf2018.L18) -> L30:
